General cleaning up:
Many json fields (eg: id,result,method) are now constants in JSONSerializer

JSONSerializer has the option of not returning the marshalled result. This can be overridden by the processed object for the serialized data

SerializerState is now an interface.

Removed Fixup information from ProcessedObject; moved this to FixupProcessedObject which extends it.